Add alternative holidays by revice related raw.
Ministry of Government Legisation gives notice[^1] that a presidential decree[^2] will be reviced for resting rights of nation and revive economy.
This revision allow additional alternative holidays for Buddha's Birthday and Christmas.
This is on progress of raw system managed by Ministry of Government Legisation and not be public yet, but that will be pass as soon as possible because current govenment of South Korea wants to make this public before next Buddah's Birthday.

[^1]: 법제처 공고번호 제2023-148호: https://www.moleg.go.kr/lawinfo/makingInfo.mo?mid=a10104010000&lawSeq=72418&lawCd=0&lawType=TYPE5&currentPage=1&keyField=&keyWord=&stYdFmt=&edYdFmt=&lsClsCd=&cptOfiOrgCd=
[^2]: 관공서의 공휴일에 관한 규정: https://www.law.go.kr/%EB%B2%95%EB%A0%B9/%EA%B4%80%EA%B3%B5%EC%84%9C%EC%9D%98%20%EA%B3%B5%ED%9C%B4%EC%9D%BC%EC%97%90%20%EA%B4%80%ED%95%9C%20%EA%B7%9C%EC%A0%95